% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_add.R
\name{afun}
\alias{afun}
\title{Estimate the coefficients of the driving factors}
\usage{
afun(k_value, kv, ku)
}
\arguments{
\item{k_value}{Mean intensity of the driving factor effect, with the default
value set to max_disp.}

\item{kv}{The data of driving factors along the x-axis.}

\item{ku}{The data of driving factors along the y-axis.}
}
\value{
a
}
\description{
Calculate the proportional coefficients of the driving factors.
}
\details{
afun function
}
