% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{dispersal}
\alias{dispersal}
\title{Using the dispersal ability of species to constrain future potential
distribution}
\usage{
dispersal(all_infor, occ, n = 8, out_dir=NULL, figure = TRUE)
}
\arguments{
\item{all_infor}{data frame. it has seven columns ( "time", "path", "thr",
"max_disp","frequency_n", "exposure_threshold", "barrier") at least.
"time": numeric, a time of the future.
"path": logical, the path saved for the predicted raster (0-1) file at a
certain time in the future.
"thr": numeric, threshold for Binaring raster files.
"max_disp": numeric, The maximum diffusion distance of a species per unit
time.
"frequency_n": numeric, the quantity per unit time during a time quantum.
"exposure_threshold": How long can species survive in unsuitable
environments.
"barrier": logical, The path of raster data for the formation of obstacles
to species diffusion. Obstacle terrain includes altitude, water system, etc.}

\item{occ}{data.frame. Species occurrence data.}

\item{n}{numeric. The number of directions a cell diffusion. the default is
4.}

\item{out_dir}{The complete path to the directory where the raster is written
as a. tif file. The diffusion results and exposure threshold results at each
time will be saved. if NULL, The result will be output in the working
directory.}

\item{figure}{logical. If TRUE, the intermediate results will plot. the
default is FALSE.}

\item{native}{logical. If TRUE, the species is native species,  the
default is TRUE.}

\item{buffer_dist}{logical. When native is FALSE, the current range of
species can be provided. If buffer_dist is NULL, the default range is 10
times the maximum distance from the occurrence points.}

\item{k_value}{logical. Mean intensity of the driving factor effect. If
k_value is NULL, the default value set to max_disp.}

\item{distribution}{Dispersal capacity distribution.}
}
\value{
Return the SpatRaster of the corresponding time within the given
storage path.
}
\description{
This function adds the diffusion ability of species to the
potential distribution of binarization. Species can diffuse at different
frequencies and speeds, and combine them with potential suitable areas in
the future.
}
\examples{
#Change to actual data frame
ALL_infor <- readRDS(system.file("extdata", "ALL_infor.rda",
                                 package = "dispersalENM"))
occourence <- readRDS(system.file("extdata", "occourence.rda",
                                 package = "dispersalENM"))
#Result storage path
dirout <- getwd()

#Copy the predicted raster files from the system
file.copy(system.file("extdata", "tif", package = "dispersalENM"), dirout,
                     recursive = TRUE)
file.copy(system.file("extdata", "barrier", package = "dispersalENM"), dirout,
                      recursive = TRUE)
#Dispersal
dispersal(all_infor = ALL_infor, occ = occourence, out_dir=dirout, n = 8,
figure = TRUE)

}
