% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_add.R
\name{distri}
\alias{distri}
\title{Diffusion distribution}
\usage{
distri(distribution)
}
\arguments{
\item{distribution}{Diffusion distribution}
}
\value{
value
}
\description{
Diffusion distribution
}
\details{
distri function
}
