% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_correlation.R
\name{env_correlation}
\alias{env_correlation}
\title{Correlation analysis of environmental data}
\usage{
env_correlation(occs, env)
}
\arguments{
\item{occs}{Data.frame. species occurrence data.}

\item{env}{SpatRaster. environmental data.}
}
\value{
vector
}
\description{
Perform correlation analysis on environmental data to avoid
multicollinearity.
}
\details{
Title
}
\examples{
occourence <- readRDS(system.file("extdata", "occourence.rda",
                                 package = "dispersalENM"))
trainenv <-  terra::rast(list.files(
  system.file("extdata", "env/trainenv", package = "dispersalENM"),
  pattern = "tif", full.names = TRUE))
data <- env_correlation(occs = occourence, env = trainenv)

}
