% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_add.R
\name{expand_buffer}
\alias{expand_buffer}
\title{Calculate the buffer of a cell after diffusion}
\usage{
expand_buffer(point, n = 8, distances = NA, max_disp, pixel, crs)
}
\arguments{
\item{point}{A point containing x and y}

\item{n}{numeric. The number of directions a cell diffusion. the default is
4.}

\item{distances}{numeric. The diffusion distance. if NA,  calculated based
on the exponential distribution and maximum diffusion distance.}

\item{max_disp}{numeric. the maximum diffusion distance.}

\item{density}{the species dispersal distance density distribution.}

\item{pixel}{numeric. the resolution of environmental data.}

\item{crs}{the points’ projection.}

\item{a}{the coefficients of the driving factors.}
}
\value{
polygon
}
\description{
After species diffuse at n angles on a certain cell, n points
are connected to form a buffer, which is the diffusion buffer of a certain
cell.
}
\details{
expand_buffer function
}
