% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_add.R
\name{get_disp}
\alias{get_disp}
\title{Calculate the diffusion distance in each direction}
\usage{
get_disp(n, max_disp, pixel, density, R)
}
\arguments{
\item{n}{numeric. The number of directions a cell diffusion. the default is
4.}

\item{max_disp}{numeric. The maximum diffusion distance of a species per unit
time.}

\item{pixel}{numeric. the resolution of environmental data.}

\item{density}{the species dispersal distance density distribution.}

\item{R}{numeric. the resistance of environmental data.}
}
\value{
vector
}
\description{
Each cell has n diffusion directions, and the diffusion distance
of species in each direction is calculated based on the maximum diffusion
distance and exponential distribution.
}
\details{
get_disp function
}
