% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_proj.R
\name{mask_proj}
\alias{mask_proj}
\title{base map}
\usage{
mask_proj(env_1)
}
\arguments{
\item{env_1}{SpatRaster. One of the environmental data.}
}
\value{
SpatRaster
}
\description{
Create a base map of continuous numbers from 0 to n.
}
\details{
mask_proj function
}
