% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_model.R
\name{maxent_model}
\alias{maxent_model}
\title{MaxEnt model}
\usage{
maxent_model(occs, env, fc = c("L", "Q", "P"),
                   nbg = 10000, rm = 1, doclamp = "false")
}
\arguments{
\item{occs}{Data.frame. species occurrence data.}

\item{env}{SpatRaster. environmental data.}

\item{fc}{Feature combination. the option including "L"(linear),
"Q"(quadratic), "P"(product), "T"(threshold), "H"(hinge). the default is
c("L","Q","P").}

\item{nbg}{Numeric. The quantity of background data. Default is 10000.}

\item{rm}{Regularization multiplier. The default is 1.}

\item{doclamp}{Logical. If "true", using Jackknife method. The default is
"false".}
}
\value{
model
}
\description{
Building the MaxEnt model.
}
\details{
MaxEntmodel function
}
\examples{
occourence <- readRDS(system.file("extdata", "occourence.rda",
                                 package = "dispersalENM"))
trainenv <-  terra::rast(list.files(
system.file("extdata", "env/trainenv", package = "dispersalENM"),
pattern = "tif", full.names = TRUE))
model <- maxent_model(occourence, trainenv, fc=c("L"), nbg=10000,
                      rm=1, doclamp="false")

}
