% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tunning.R
\name{model_tuning}
\alias{model_tuning}
\title{Select the optimal parameter among the alternative parameters of
MaxEnt model}
\usage{
model_tuning(occs,
             env,
             nbg = 10000,
             f_c = c("L", "Q", "P"),
             r_m = c(1, 1, 3),
             train_ratio = 0.7)
}
\arguments{
\item{occs}{data.frame. species occurrence data.}

\item{env}{SpatRaster. environmental data.}

\item{nbg}{numeric. The quantity of background data. Default is 10000.}

\item{f_c}{feature combination. The option including "L"(linear),
"Q"(quadratic), "P"(product), "T"(threshold), "H"(hinge). Combine these
selected options and optimize the model. the default is c("L","Q","P")}

\item{r_m}{regularization multiplier. The format like default c(1, 1, 3).
The first number is the starting value, the third number is the termination
value, and the second number is step.}

\item{train_ratio}{0-1. Proportion of training data}
}
\value{
a data frame
}
\description{
In the Maxent model, regularization multiplier and feature
combination are two important parameters. By model tuning, the optimal
combination is calculated after a given set of alternative parameter
combination.
}
\details{
model_tuning function
}
\examples{
occourence <- readRDS(system.file("extdata", "occourence.rda",
                                 package = "dispersalENM"))
trainenv <-  terra::rast(list.files(
system.file("extdata", "env/trainenv", package = "dispersalENM"),
pattern = "tif", full.names = TRUE))
#model_tuning
model_tuning_results <- model_tuning(occourence, trainenv, nbg=10000,
                                       f_c=c("L","Q"), r_m=c(1, 1, 3))
}
