% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_filter.R
\name{occ_filter}
\alias{occ_filter}
\title{Species occurrence data filtering}
\usage{
occ_filter(occs, env_1)
}
\arguments{
\item{occs}{Data.frame. species occurrence data.}

\item{env_1}{SpatRaster. One of the environmental data.}
}
\value{
data frame
}
\description{
To avoid bias, process species occurrence data. 1. Remove data
that is missing x or y; 2. Only one duplicate data is retained; 3. Only one
piece of data is retained within each grid.
}
\details{
occfilter function
}
\examples{
occ_data <- readRDS(system.file("extdata", "occ_data.rda",
package = "dispersalENM"))
trainenv <-  terra::rast(list.files(
  system.file("extdata", "env/trainenv", package = "dispersalENM"),
  pattern = "tif", full.names = TRUE))
occourence <-occ_filter(occs = occ_data[,c("lon","lat")],
                        env_1 = trainenv[[1]])

}
