% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{Prediction}
\usage{
prediction(model, trainenv, vars, pre_infor, dir_out, Format=".tif$")
}
\arguments{
\item{model}{ecological niche model}

\item{trainenv}{SpatRaster. The environmental data which is used in model.}

\item{vars}{Environment variables are identified in the environment data.}

\item{pre_infor}{data frame. the information of future environmental data.}

\item{dir_out}{Path of results.}

\item{Format}{Format of climate data is GeoTIFF(.tif).}
}
\value{
Return the SpatRaster of the corresponding time within the given
storage path.
}
\description{
Predict based on models and climate data.
}
\details{
Prediction function
}
\examples{
\dontrun{
dir_out <- getwd()
model <- readRDS(system.file("extdata", "model.rda",
                                 package = "dispersalENM"))
pre_infor <- readRDS(system.file("extdata", "pre_infor.rda",
                                 package = "dispersalENM"))
trainenv <- terra::rast(list.files(
system.file("extdata", "env/trainenv", package = "dispersalENM"),
pattern = "tif", full.names = TRUE))
file.copy(system.file("extdata", "env/predictenv", package = "dispersalENM"), dir_out,
                     recursive = TRUE)
vars <- c("tasmin", "tasmax", "pre")
prediction(model, trainenv, vars, pre_infor, dir_out)
}
}
