% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{statistics}
\alias{statistics}
\title{Statistical analysis of the results}
\usage{
statistics(folder_path, out_dir)
}
\arguments{
\item{folder_path}{species dispersal file("tif" format) path.}

\item{out_dir}{The complete path to the directory where the raster is written
as a. tif file. The diffusion results and exposure threshold results at each
time will be saved. if NULL, The result will be output in the working
directory.}
}
\value{
Return the SpatRaster of the corresponding time within the given
storage path.
}
\description{
Calculate the ratio of potential distribution area (compared to
the study area) within the time interval, gain, refuge, and loss of
potential distribution area.
}
\examples{
#Change to actual data frame
folder_path <- sprintf("\%s/\%s", getwd(), "dispersal_tif")
out_dir <- getwd()
statistics(folder_path = folder_path, out_dir = out_dir)

}
